requirements.txt
matplotlib==3.7.1
numba==0.57.1
numpy==1.23.5
open3d==0.17.0
opencv_python==4.7.0.68
Pillow==9.3.0
Pillow==10.3.0
progressbar33==2.4
pycuda==2024.1
pyhocon==0.3.57
PyMCubes==0.1.4
setuptools==68.2.2
setuptools==60.10.0
skimage==0.0
tensorboardX==2.6
tensorboardX==2.6.2.2
torch==1.13.1
torchvision==0.14.1
trimesh==3.22.5
